Move Name tests into a shared folder
